ユーザー注文実装完了

+++ db/migrate/20230723123534_create_orders.rb
@@ -5,7 +5,7 @@ class CreateOrders < ActiveRecord::Migration[6.1]
       t.string :shipping_address, null: false
       t.string :shipping_postal_code, null: false
       t.string :shipping_name, null: false
-      t.decimal :shipping_fee, null: false
+      t.integer :shipping_fee, null: false
       t.integer :payment_method, null: false, default: 0
       t.decimal :total_amount, null: false
